Elliptic and rectangular mask views are inheriting from common base.
