+++ GUI/coregui/Views/MaskWidgets/RectangleView.h
@@ -17,16 +17,12 @@
 #ifndef RECTANGLEVIEW_H
 #define RECTANGLEVIEW_H
 
-#include "IMaskView.h"
-#include "SizeHandleElement.h"
-#include <QMap>
+#include "RectangleBaseView.h"
 
-//! This is a View of rectangular mask ( represented by RectangleItem) on GraphicsScene.
+//! This is a View of rectangular mask (represented by RectangleItem) on GraphicsScene.
 //! Given view follows standard QGraphicsScene notations: (x,y) is top left corner.
-//!
-//! FIXME Refactor RectangleView and EllipseView to rely on common base class
 
-class BA_CORE_API_ RectangleView : public IMaskView
+class BA_CORE_API_ RectangleView : public RectangleBaseView
 {
     Q_OBJECT
 
@@ -40,35 +36,17 @@ public slots:
     virtual void onChangedY();
     virtual void onPropertyChange(const QString &propertyName);
 
-private slots:
-    void onSizeHandleElementRequest(bool going_to_resize);
-
 protected:
     void paint(QPainter *painter, const QStyleOptionGraphicsItem *, QWidget *);
-    QVariant itemChange(GraphicsItemChange change, const QVariant &value);
-
     void mouseMoveEvent(QGraphicsSceneMouseEvent *event);
-    void mouseReleaseEvent(QGraphicsSceneMouseEvent *event);
 
 protected:
-    void update_view();
-    void update_bounding_rect();
     void update_position();
+    QRectF mask_rectangle();
     qreal left() const;
     qreal right() const;
-    qreal width() const;
     qreal top() const;
     qreal bottom() const;
-    qreal height() const;
-    void create_size_handle_elements();
-    bool m_block_on_property_change;
-    QRectF m_mask_rect; // mask rectangle in scene coordinates
-    QMap<SizeHandleElement::EHandleLocation, SizeHandleElement *> m_resize_handles;
-     //!< coordinates of corner opposite to the grip corner at the moment it first clicked
-     //!< in scene coordinates
-    QPointF m_resize_opposite_origin;
-    SizeHandleElement *m_activeHandleElement;
 };
-
 
 #endif // RECTANGLEVIEW_H
